Update
commented state script. Still not sure on how to display enemy path. I have a better understanding of state switching and how to call and update enemy path(CalculatePath). I know where the position for enemy, spawner, base, and walls are set but am unclear on how to alter/change it so that I can have them appear above the floor.