Update 2
same situation as the previous update, but I did add a few more comments. 

random thought: not sure why I decided to not include failure screen after going through the code because I believe that It will cause an error or will crash maybe from not knowing what to do in failure phase. might add that in later